:recycle: Integrate EmptyActivityConverter into EmptyActivityService
