+++ src/converters/index.ts
@@ -1,4 +1,3 @@
-export * from './empty_activity_converter';
 export * from './event_converter';
 export * from './user_task_converter';
 export * from './manual_task_converter';
