+++ src/empty_activity_service.ts
@@ -3,42 +3,55 @@ import {IEventAggregator, Subscription} from '@essential-projects/event_aggregat
 import {IIAMService, IIdentity} from '@essential-projects/iam_contracts';
 import {APIs, DataModels, Messages} from '@process-engine/consumer_api_contracts';
 import {
+  BpmnType,
   FlowNodeInstance,
   FlowNodeInstanceState,
+  ICorrelationService,
   IFlowNodeInstanceService,
+  IProcessModelUseCases,
+  Model,
+  ProcessTokenType,
 } from '@process-engine/persistence_api.contracts';
-import {FinishEmptyActivityMessage as InternalFinishEmptyActivityMessage} from '@process-engine/process_engine_contracts';
+import {
+  IProcessModelFacade,
+  IProcessModelFacadeFactory,
+  FinishEmptyActivityMessage as InternalFinishEmptyActivityMessage,
+} from '@process-engine/process_engine_contracts';
 
 import {NotificationAdapter} from './adapters/index';
-import {EmptyActivityConverter} from './converters/index';
 import {applyPagination} from './paginator';
+
+import * as ProcessModelCache from './converters/process_model_cache';
 
 export class EmptyActivityService implements APIs.IEmptyActivityConsumerApi {
 
+  private readonly correlationService: ICorrelationService;
   private readonly eventAggregator: IEventAggregator;
   private readonly flowNodeInstanceService: IFlowNodeInstanceService;
   private readonly iamService: IIAMService;
+  private readonly processModelUseCase: IProcessModelUseCases;
+  private readonly processModelFacadeFactory: IProcessModelFacadeFactory;
 
   private readonly notificationAdapter: NotificationAdapter;
-
-  private readonly emptyActivityConverter: EmptyActivityConverter;
 
   private readonly canSubscribeToEventsClaim = 'can_subscribe_to_events';
 
   constructor(
+    correlationService: ICorrelationService,
     eventAggregator: IEventAggregator,
     flowNodeInstanceService: IFlowNodeInstanceService,
     iamService: IIAMService,
     notificationAdapter: NotificationAdapter,
-    emptyActivityConverter: EmptyActivityConverter,
+    processModelFacadeFactory: IProcessModelFacadeFactory,
+    processModelUseCase: IProcessModelUseCases,
   ) {
+    this.correlationService = correlationService;
     this.eventAggregator = eventAggregator;
     this.flowNodeInstanceService = flowNodeInstanceService;
     this.iamService = iamService;
-
     this.notificationAdapter = notificationAdapter;
-
-    this.emptyActivityConverter = emptyActivityConverter;
+    this.processModelFacadeFactory = processModelFacadeFactory;
+    this.processModelUseCase = processModelUseCase;
   }
 
   public async onEmptyActivityWaiting(
@@ -90,7 +103,9 @@ export class EmptyActivityService implements APIs.IEmptyActivityConsumerApi {
 
     const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByProcessModel(processModelId);
 
-    const emptyActivityList = await this.emptyActivityConverter.convert(identity, suspendedFlowNodes);
+    const emptyActivities = suspendedFlowNodes.filter(this.checkIfIsFlowNodeIsEmptyActivity);
+
+    const emptyActivityList = await this.convertToEmptyActivity(identity, emptyActivities);
 
     // TODO: Remove that useless `EmptyActivityList` datatype and just return an Array of EmptyActivities.
     // Goes for the other UseCases as well.
@@ -108,7 +123,9 @@ export class EmptyActivityService implements APIs.IEmptyActivityConsumerApi {
 
     const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByProcessInstance(processInstanceId);
 
-    const emptyActivityList = await this.emptyActivityConverter.convert(identity, suspendedFlowNodes);
+    const emptyActivities = suspendedFlowNodes.filter(this.checkIfIsFlowNodeIsEmptyActivity);
+
+    const emptyActivityList = await this.convertToEmptyActivity(identity, emptyActivities);
 
     emptyActivityList.emptyActivities = applyPagination(emptyActivityList.emptyActivities, offset, limit);
 
@@ -124,7 +141,9 @@ export class EmptyActivityService implements APIs.IEmptyActivityConsumerApi {
 
     const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByCorrelation(correlationId);
 
-    const emptyActivityList = await this.emptyActivityConverter.convert(identity, suspendedFlowNodes);
+    const emptyActivities = suspendedFlowNodes.filter(this.checkIfIsFlowNodeIsEmptyActivity);
+
+    const emptyActivityList = await this.convertToEmptyActivity(identity, emptyActivities);
 
     emptyActivityList.emptyActivities = applyPagination(emptyActivityList.emptyActivities, offset, limit);
 
@@ -142,10 +161,12 @@ export class EmptyActivityService implements APIs.IEmptyActivityConsumerApi {
     const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByCorrelation(correlationId);
 
     const suspendedProcessModelFlowNodes = suspendedFlowNodes.filter((flowNodeInstance: FlowNodeInstance): boolean => {
-      return flowNodeInstance.tokens[0].processModelId === processModelId;
+      const isEmptyActivity = this.checkIfIsFlowNodeIsEmptyActivity(flowNodeInstance);
+      const belongsToProcessModel = flowNodeInstance.tokens[0].processModelId === processModelId;
+      return isEmptyActivity && belongsToProcessModel;
     });
 
-    const emptyActivityList = await this.emptyActivityConverter.convert(identity, suspendedProcessModelFlowNodes);
+    const emptyActivityList = await this.convertToEmptyActivity(identity, suspendedProcessModelFlowNodes);
 
     emptyActivityList.emptyActivities = applyPagination(emptyActivityList.emptyActivities, offset, limit);
 
@@ -161,10 +182,12 @@ export class EmptyActivityService implements APIs.IEmptyActivityConsumerApi {
     const suspendedFlowNodeInstances = await this.flowNodeInstanceService.queryByState(FlowNodeInstanceState.suspended);
 
     const flowNodeInstancesOwnedByUser = suspendedFlowNodeInstances.filter((flowNodeInstance: FlowNodeInstance): boolean => {
-      return this.checkIfIdentityUserIDsMatch(identity, flowNodeInstance.owner);
+      const isEmptyActivity = this.checkIfIsFlowNodeIsEmptyActivity(flowNodeInstance);
+      const userIdsMatch = this.checkIfIdentityUserIDsMatch(identity, flowNodeInstance.owner);
+      return isEmptyActivity && userIdsMatch;
     });
 
-    const emptyActivityList = await this.emptyActivityConverter.convert(identity, flowNodeInstancesOwnedByUser);
+    const emptyActivityList = await this.convertToEmptyActivity(identity, flowNodeInstancesOwnedByUser);
 
     emptyActivityList.emptyActivities = applyPagination(emptyActivityList.emptyActivities, offset, limit);
 
@@ -181,15 +204,14 @@ export class EmptyActivityService implements APIs.IEmptyActivityConsumerApi {
     const matchingFlowNodeInstance =
       await this.getFlowNodeInstanceForCorrelationInProcessInstance(correlationId, processInstanceId, emptyActivityInstanceId);
 
-    const noMatchingInstanceFound = matchingFlowNodeInstance === undefined;
-    if (noMatchingInstanceFound) {
+    if (matchingFlowNodeInstance === undefined) {
       const errorMessage =
         // eslint-disable-next-line max-len
         `ProcessInstance '${processInstanceId}' in Correlation '${correlationId}' does not have an EmptyActivity with id '${emptyActivityInstanceId}'`;
       throw new EssentialProjectErrors.NotFoundError(errorMessage);
     }
 
-    const convertedEmptyActivityList = await this.emptyActivityConverter.convert(identity, [matchingFlowNodeInstance]);
+    const convertedEmptyActivityList = await this.convertToEmptyActivity(identity, [matchingFlowNodeInstance]);
 
     const matchingEmptyActivity = convertedEmptyActivityList.emptyActivities[0];
 
@@ -226,8 +248,90 @@ export class EmptyActivityService implements APIs.IEmptyActivityConsumerApi {
     return matchingInstance;
   }
 
+  private checkIfIsFlowNodeIsEmptyActivity(flowNodeInstance: FlowNodeInstance): boolean {
+    return flowNodeInstance.flowNodeType !== BpmnType.emptyActivity;
+  }
+
   private checkIfIdentityUserIDsMatch(identityA: IIdentity, identityB: IIdentity): boolean {
     return identityA.userId === identityB.userId;
+  }
+
+  private async convertToEmptyActivity(
+    identity: IIdentity,
+    suspendedFlowNodes: Array<FlowNodeInstance>,
+  ): Promise<DataModels.EmptyActivities.EmptyActivityList> {
+
+    const suspendedEmptyActivities: Array<DataModels.EmptyActivities.EmptyActivity> = [];
+
+    for (const suspendedFlowNode of suspendedFlowNodes) {
+      const processModelFacade = await this.getProcessModelForFlowNodeInstance(identity, suspendedFlowNode);
+
+      const emptyActivity = await this.convertSuspendedFlowNodeToEmptyActivity(suspendedFlowNode, processModelFacade);
+
+      suspendedEmptyActivities.push(emptyActivity);
+    }
+
+    const emptyActivityList: DataModels.EmptyActivities.EmptyActivityList = {
+      emptyActivities: suspendedEmptyActivities,
+      totalCount: suspendedEmptyActivities.length,
+    };
+
+    return emptyActivityList;
+  }
+
+  private async getProcessModelForFlowNodeInstance(
+    identity: IIdentity,
+    flowNodeInstance: FlowNodeInstance,
+  ): Promise<IProcessModelFacade> {
+
+    let processModel: Model.Process;
+
+    // We must store the ProcessModel for each user, to account for lane-restrictions.
+    // Some users may not be able to see some lanes that are visible to others.
+    const cacheKeyToUse = `${flowNodeInstance.processInstanceId}-${identity.userId}`;
+
+    const cacheHasMatchingEntry = ProcessModelCache.hasEntry(cacheKeyToUse);
+    if (cacheHasMatchingEntry) {
+      processModel = ProcessModelCache.get(cacheKeyToUse);
+    } else {
+      const processModelHash = await this.getProcessModelHashForProcessInstance(identity, flowNodeInstance.processInstanceId);
+      processModel = await this.processModelUseCase.getByHash(identity, flowNodeInstance.processModelId, processModelHash);
+      ProcessModelCache.add(cacheKeyToUse, processModel);
+    }
+
+    const processModelFacade = this.processModelFacadeFactory.create(processModel);
+
+    return processModelFacade;
+  }
+
+  private async getProcessModelHashForProcessInstance(identity: IIdentity, processInstanceId: string): Promise<string> {
+    const processInstance = await this.correlationService.getByProcessInstanceId(identity, processInstanceId);
+
+    return processInstance.hash;
+  }
+
+  private async convertSuspendedFlowNodeToEmptyActivity(
+    emptyActivityInstance: FlowNodeInstance,
+    processModelFacade: IProcessModelFacade,
+  ): Promise<DataModels.EmptyActivities.EmptyActivity> {
+
+    const emptyActivityModel = processModelFacade.getFlowNodeById(emptyActivityInstance.flowNodeId);
+
+    const onSuspendToken = emptyActivityInstance.getTokenByType(ProcessTokenType.onSuspend);
+
+    const consumerApiEmptyActivity: DataModels.EmptyActivities.EmptyActivity = {
+      flowNodeType: BpmnType.emptyActivity,
+      id: emptyActivityInstance.flowNodeId,
+      flowNodeInstanceId: emptyActivityInstance.id,
+      name: emptyActivityModel.name,
+      correlationId: emptyActivityInstance.correlationId,
+      processModelId: emptyActivityInstance.processModelId,
+      processInstanceId: emptyActivityInstance.processInstanceId,
+      tokenPayload: onSuspendToken.payload,
+    };
+
+    return consumerApiEmptyActivity;
+
   }
 
   private publishFinishEmptyActivityEvent(
